requirements.txt
beautifulsoup4==4.12.3
instructor==1.3.7
openai==1.37.1
pydantic==2.8.2
pytest-playwright==0.5.1
python-dotenv==1.0.1
